+++ client/src/main/java/io/lionweb/client/delta/messages/commands/references/MoveEntryInSameReference.java
@@ -8,13 +8,13 @@ import org.jetbrains.annotations.Nullable;
 
 /**
  * Move existing entry movedTarget/movedResolveInfo inside parent's reference at oldIndex inside
- * parent's reference at newIndex.
+ * parent's reference at indexOffset.
  */
 public final class MoveEntryInSameReference extends DeltaCommand {
   public final @NotNull String parent;
   public final @NotNull MetaPointer reference;
   public final int oldIndex;
-  public final int newIndex;
+  public final int indexOffset;
   public final @Nullable String movedTarget;
   public final @Nullable String movedResolveInfo;
 
@@ -23,7 +23,7 @@ public final class MoveEntryInSameReference extends DeltaCommand {
       @Nullable String parent,
       @NotNull MetaPointer reference,
       int oldIndex,
-      int newIndex,
+      int indexOffset,
       @Nullable String movedTarget,
       @Nullable String movedResolveInfo) {
     super(commandId);
@@ -32,13 +32,10 @@ public final class MoveEntryInSameReference extends DeltaCommand {
     if (oldIndex < 0) {
       throw new IllegalArgumentException("oldIndex must be non-negative");
     }
-    if (newIndex < 0) {
-      throw new IllegalArgumentException("newIndex must be non-negative");
-    }
     this.parent = parent;
     this.reference = reference;
     this.oldIndex = oldIndex;
-    this.newIndex = newIndex;
+    this.indexOffset = indexOffset;
     this.movedTarget = movedTarget;
     this.movedResolveInfo = movedResolveInfo;
   }
@@ -53,8 +50,8 @@ public final class MoveEntryInSameReference extends DeltaCommand {
         + reference
         + ", oldIndex="
         + oldIndex
-        + ", newIndex="
-        + newIndex
+        + ", indexOffset="
+        + indexOffset
         + ", movedTarget='"
         + movedTarget
         + '\''