+++ client/src/main/java/io/lionweb/client/delta/messages/events/annotations/AnnotationMovedInSameParent.java
@@ -6,30 +6,27 @@ import org.jetbrains.annotations.NotNull;
 
 /**
  * Existing node movedAnnotation (previously inside parent's annotations at oldIndex) has been moved
- * inside parent's annotations at newIndex.
+ * inside parent's annotations at indexOffset.
  */
 public class AnnotationMovedInSameParent extends BaseDeltaEvent<AnnotationMovedInSameParent> {
-  public final int newIndex;
+  public final int indexOffset;
   public final @NotNull String movedAnnotation;
   public final @NotNull String parent;
   public final int oldIndex;
 
   public AnnotationMovedInSameParent(
       int sequenceNumber,
-      int newIndex,
+      int indexOffset,
       @NotNull String movedAnnotation,
       @NotNull String parent,
       int oldIndex) {
     super(sequenceNumber);
     Objects.requireNonNull(movedAnnotation, "movedAnnotation should not be null");
     Objects.requireNonNull(parent, "parent should not be null");
-    if (newIndex < 0) {
-      throw new IllegalArgumentException("newIndex should be non-negative");
-    }
     if (oldIndex < 0) {
       throw new IllegalArgumentException("oldIndex should be non-negative");
     }
-    this.newIndex = newIndex;
+    this.indexOffset = indexOffset;
     this.movedAnnotation = movedAnnotation;
     this.parent = parent;
     this.oldIndex = oldIndex;
@@ -38,8 +35,8 @@ public class AnnotationMovedInSameParent extends BaseDeltaEvent<AnnotationMovedI
   @Override
   public String toString() {
     return "AnnotationMovedInSameParent{"
-        + "newIndex="
-        + newIndex
+        + "indexOffset="
+        + indexOffset
         + ", movedAnnotation='"
         + movedAnnotation
         + '\''