+++ client/src/main/java/io/lionweb/client/inmemory/DeltaCommandReceiverImpl.java
@@ -221,12 +221,13 @@ class DeltaCommandReceiverImpl implements DeltaCommandReceiver {
     SerializedClassifierInstance parent = requireNode(data, cmd.parent);
     List<String> annotations = new ArrayList<>(parent.getAnnotations());
     String id = annotations.remove(cmd.oldIndex);
-    annotations.add(cmd.newIndex, id);
+    int newIndex = cmd.oldIndex + cmd.indexOffset;
+    annotations.add(newIndex, id);
     parent.setAnnotations(annotations);
     channel.sendEvent(
         seqNum ->
             new AnnotationMovedInSameParent(
-                    seqNum, cmd.newIndex, cmd.movedAnnotation, cmd.parent, cmd.oldIndex)
+                    seqNum, cmd.indexOffset, cmd.movedAnnotation, cmd.parent, cmd.oldIndex)
                 .addSource(source));
   }
 
@@ -296,11 +297,17 @@ class DeltaCommandReceiverImpl implements DeltaCommandReceiver {
       @NotNull CommandSource source) {
     SerializedClassifierInstance parent = requireNode(data, cmd.parent);
     parent.removeChild(cmd.movedChild);
-    parent.addChild(cmd.containment, cmd.movedChild, cmd.newIndex);
+    int newIndex = cmd.oldIndex + cmd.indexOffset;
+    parent.addChild(cmd.containment, cmd.movedChild, newIndex);
     channel.sendEvent(
         seqNum ->
             new ChildMovedInSameContainment(
-                    seqNum, cmd.newIndex, cmd.movedChild, cmd.parent, cmd.containment, cmd.oldIndex)
+                    seqNum,
+                    cmd.indexOffset,
+                    cmd.movedChild,
+                    cmd.parent,
+                    cmd.containment,
+                    cmd.oldIndex)
                 .addSource(source));
   }